+++ pymilo/pymilo_param.py
@@ -210,7 +210,7 @@ SKLEARN_PREPROCESSING_TABLE = {
     "Normalizer": preprocessing.Normalizer,
     "OrdinalEncoder": preprocessing.OrdinalEncoder,
     "PolynomialFeatures": preprocessing.PolynomialFeatures,
-    "PowerTransformer": preprocessing.PowerTransformer,
+    "RobustScaler": preprocessing.RobustScaler,
 }
 
 KEYS_NEED_PREPROCESSING_BEFORE_DESERIALIZATION = {