+++ tests/test_preprocessings/test_preprocessings.py
@@ -13,7 +13,7 @@ from max_abs_scaler import max_abs_scaler
 from normalizer import normalizer
 from ordinal_encoder import ordinal_encoder
 from polynomial_features import polynomial_features
-from standard_scaler import standard_scaler
+from robust_scaler import robust_scaler
 
 PREPROCESSINGS = [one_hot_encoder,
                   label_binarizer,
@@ -27,7 +27,7 @@ PREPROCESSINGS = [one_hot_encoder,
                   normalizer,
                   ordinal_encoder,
                   polynomial_features,
-                  standard_scaler
+                  robust_scaler,
                   ]
 
 @pytest.fixture(scope="session", autouse=True)